requirement.txt
langchain
langchain-community
langchain-dotenv
pydantic
